+++ core/dbt/parser/search.py
@@ -63,31 +63,24 @@ class FullBlock(FileBlock):
         return self.block.full_block
 
 
-class FilesystemSearcher(Iterable[FilePath]):
-    def __init__(
-        self, project: Project, relative_dirs: List[str], extension: str
-    ) -> None:
-        self.project = project
-        self.relative_dirs = relative_dirs
-        self.extension = extension
-
-    def __iter__(self) -> Iterator[FilePath]:
-        ext = "[!.#~]*" + self.extension
-
-        root = self.project.project_root
-
-        for result in find_matching(root, self.relative_dirs, ext):
-            if 'searched_path' not in result or 'relative_path' not in result:
-                raise InternalException(
-                    'Invalid result from find_matching: {}'.format(result)
-                )
-            file_match = FilePath(
-                searched_path=result['searched_path'],
-                relative_path=result['relative_path'],
-                modification_time=result['modification_time'],
-                project_root=root,
+def filesystem_search(project: Project, relative_dirs: List[str], extension: str):
+    ext = "[!.#~]*" + extension
+    root = project.project_root
+    file_path_list = []
+    for result in find_matching(root, relative_dirs, ext):
+        if 'searched_path' not in result or 'relative_path' not in result:
+            raise InternalException(
+                'Invalid result from find_matching: {}'.format(result)
             )
-            yield file_match
+        file_match = FilePath(
+            searched_path=result['searched_path'],
+            relative_path=result['relative_path'],
+            modification_time=result['modification_time'],
+            project_root=root,
+        )
+        file_path_list.append(file_match)
+
+    return file_path_list
 
 
 Block = Union[BlockContents, FullBlock]